requirements.txt
autopep8==1.4.4
firebase-admin==3.2.1
Flask==1.1.1
pylint==2.3.1
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
python-firebase==1.2
python-telegram-bot==12.3.0
requests==2.22.0
selenium==3.141.0
six==1.12.0
tornado==6.0.3
waitress==1.4.3
